Keep window view and undo history on reload
The previous behavior was to just save the cursor position. That worked
fine in most cases, but caused the window to scroll back up when it was
scrolled more than the height of the text. Info about winsaveview is
available on ':help winsaveview()'. The undo history is kept by
temporarily turning on the 'hidden' option. That prevents the code
buffer from being abandoned when the buffer switch is made.