Update README.md
With Vim 8 or later you don't need to use a plug-in manager like Pathogen or alike. I couldn't get the interactive editing commands working with the instructions in the this README but following the steps that i wrote it worked, so i think that this might be helpful for others.